Fixed bugs

+++ src/utilities.c
@@ -479,9 +479,9 @@ void plotSolution(Instance *d, const char * plotPixelSize, const char * pointCol
     fprintf(gnuplotPipe, "e\n");
     
     // second print the tour
-    for (size_t i = 0; i < d->nodesCount; i++)
+    for (int i = 0; i < d->nodesCount; i++)
     {
-        size_t successorID = d->solution.bestSolution[i];
+        int successorID = d->solution.bestSolution[i];
         fprintf(gnuplotPipe, "%f %f\n", d->X[successorID], d->Y[successorID]);
     }
     fprintf(gnuplotPipe, "%lf %lf\n", d->X[d->solution.bestSolution[0]], d->Y[d->solution.bestSolution[0]]);
